server/requirements.txt
aiohttp==3.8.5
python-socketio==5.8.0
mido==1.3.0
python-rtmidi==1.5.5
websocket-client==1.6.1
requests==2.31.0
